requirements.txt
absl-py==2.0.0
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.44.0
fsspec==2023.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
idna==3.4
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.21.2
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
Werkzeug==3.0.1
Wikipedia-API==0.6.0
